requirements.txt
langchain>=0.1.16
langchain-core>=0.1.35
langchain-community>=0.0.26
langchain-groq>=0.1.4
google-generativeai>=0.3.0
langchain-google-genai
python-dotenv
requests
# diffusers
# torch
# transformers
# accelerate
playwright